requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.22
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector-python==8.0.23
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
protobuf==3.15.3
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
twilio==6.53.0
urllib3==1.26.3
